pyproject.toml
[tool.poetry]
name = "weeekly-meeting"
version = "0.1.0"
description = "WEBUI for Weekly Meeting"
authors = ["Murali Krishnan Raajanthira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
streamlit = "^1.31.1"
python-pptx = "^0.6.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
